update ddp test
