+++ tests/tests_fabric/strategies/test_ddp.py
@@ -27,7 +27,6 @@ from lightning.fabric.utilities.imports import _TORCH_GREATER_EQUAL_2_0
 from torch.nn.parallel import DistributedDataParallel
 
 from tests_fabric.helpers.runif import RunIf
-from tests_fabric.strategies.test_single_device import _MyFabricGradNorm, _MyFabricGradVal
 
 
 @pytest.mark.parametrize(
@@ -111,28 +110,6 @@ def test_ddp_module_state_dict():
     assert retrieved_state_dict.keys() == original_state_dict.keys()
     strategy.load_module_state_dict(wrapped_module, retrieved_state_dict)
     strategy.load_module_state_dict(wrapped_module, original_state_dict)
-
-
-@pytest.mark.parametrize(
-    ("clip_type", "accelerator", "precision"),
-    [
-        ("norm", "cpu", "32-true"),
-        ("val", "cpu", "32-true"),
-        ("norm", "cpu", "bf16-mixed"),
-        ("val", "cpu", "bf16-mixed"),
-        pytest.param("norm", "cuda", "32-true", marks=RunIf(min_cuda_gpus=2)),
-        pytest.param("val", "cuda", "32-true", marks=RunIf(min_cuda_gpus=2)),
-        pytest.param("norm", "cuda", "16-mixed", marks=RunIf(min_cuda_gpus=2)),
-        pytest.param("val", "cuda", "16-mixed", marks=RunIf(min_cuda_gpus=2)),
-        pytest.param("norm", "cuda", "bf16-mixed", marks=RunIf(min_cuda_gpus=2, bf16_cuda=True)),
-        pytest.param("val", "cuda", "bf16-mixed", marks=RunIf(min_cuda_gpus=2, bf16_cuda=True)),
-    ],
-)
-@RunIf(standalone=True)
-def test_ddp_grad_clipping(clip_type, accelerator, precision):
-    clipping_test_cls = _MyFabricGradNorm if clip_type == "norm" else _MyFabricGradVal
-    fabric = clipping_test_cls(accelerator=accelerator, devices=2, precision=precision, strategy="ddp")
-    fabric.run()
 
 
 @RunIf(min_cuda_gpus=2)
