+++ tests/tests_fabric/strategies/test_ddp_integration.py
@@ -18,6 +18,7 @@ import torch
 from lightning.fabric import Fabric
 
 from tests_fabric.helpers.runif import RunIf
+from tests_fabric.strategies.test_single_device import _run_test_clip_gradients
 
 
 @pytest.mark.parametrize(
@@ -61,3 +62,28 @@ def _run_ddp_save_load(fabric, tmp_path):
     assert_params_equal(params_before, wrapped_model.parameters())
     fabric.load(tmp_path / "saved_after_setup.ckpt", {"model": wrapped_model})
     assert_params_equal(params_before, wrapped_model.parameters())
+
+
+@pytest.mark.parametrize(
+    ("clip_type", "accelerator", "precision"),
+    [
+        ("norm", "cpu", "32-true"),
+        ("val", "cpu", "32-true"),
+        ("norm", "cpu", "bf16-mixed"),
+        ("val", "cpu", "bf16-mixed"),
+        pytest.param("norm", "cuda", "32-true", marks=RunIf(min_cuda_gpus=2)),
+        pytest.param("val", "cuda", "32-true", marks=RunIf(min_cuda_gpus=2)),
+        pytest.param("norm", "cuda", "16-mixed", marks=RunIf(min_cuda_gpus=2)),
+        pytest.param("val", "cuda", "16-mixed", marks=RunIf(min_cuda_gpus=2)),
+        pytest.param("norm", "cuda", "bf16-mixed", marks=RunIf(min_cuda_gpus=2, bf16_cuda=True)),
+        pytest.param("val", "cuda", "bf16-mixed", marks=RunIf(min_cuda_gpus=2, bf16_cuda=True)),
+    ],
+)
+@RunIf(standalone=True)
+def test_ddp_grad_clipping(clip_type, accelerator, precision):
+    if clip_type == "norm" and precision == "16-mixed":
+        pytest.skip(reason="Clipping by norm with 16-mixed is numerically unstable.")
+
+    fabric = Fabric(accelerator=accelerator, devices=2, precision=precision, strategy="ddp")
+    fabric.launch()
+    _run_test_clip_gradients(fabric=fabric, clip_type=clip_type)
